Refactor filterResults function and cleanup some component naming syntax
