+++ src/components/Favorites/index.js
